+++ src/App.tsx
@@ -3,13 +3,16 @@ import MobileLayout from './components/common/MobileLayout';
 import reset from './global/style/reset';
 import { theme } from './global/style/theme';
 import ShopPage from './pages/shop/ShopPage';
+import CartProvider from './context/CartContext';
 
 function App() {
   return (
     <MobileLayout>
       <Global styles={reset} />
       <ThemeProvider theme={theme}>
-        <ShopPage />
+        <CartProvider>
+          <ShopPage />
+        </CartProvider>
       </ThemeProvider>
     </MobileLayout>
   );
